requirements.txt
tornado
docopt
pyyaml
pymongo
